+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 ## 1.0.0-beta.1 - 2018-04-26
 ### Fixed
 - Class type casing issues
+- Issue where a correct cache instance would throw an exception
 
 ### Added
 - SObject field type supports min and max values